requirements.txt
google-generativeai
flask
flask-cors
python-dotenv
requests